Detect invalid ABIs on methods and throw an error.
Trying to call methods with a non-Rust ABI on a trait object results in
an ICE. While we could try to translate them, it makes more sense to
just reject non-Rust ABIs in type-check.

This is technically a [breaking-change] due to that fact that we didn't
ICE when statically dispatching these method calls.

Fixes #26049